requirements.txt
django==3.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-filter==2.4.0
gunicorn==20.1.0
psycopg2==2.9